Added support for later versions of the python bindings of libxslt which know about saving to a string (rather than the ol' tempfile hack).
--HG--
extra : convert_revision : svn%3A01e103ed-f575-20b1-d296-6440ccab0075/DataTemplates/trunk%40109

+++ plugins/plugin_libxslt.py
@@ -1,4 +1,13 @@
-import libxslt, libxml2, StringIO
+import libxslt, libxml2, StringIO, sys
+
+if hasattr(libxslt.stylesheet, 'saveResultToString'):
+    RENDERVIAFILE=0
+else:
+    RENDERVIAFILE=1
+    sys.stderr.write('The libxslt plugin for the DataTemplate product '
+                     'is using an old version of libxslt -- consider upgrading '
+                     'to the latest version to avoid workarounds\n')
+    sys.stderr.flush()
 
 class UriResolver:                                                                   
     def __init__(self,context=None):
@@ -33,25 +42,35 @@ def render(self, source_xml, content_type):
     """
     import tempfile, os, StringIO
 
+    # handle URI's as if they were in the ZODB
     resolver = UriResolver(self).resolve
     libxml2.setEntityLoader(resolver)
+    
     styledoc = libxml2.parseDoc(self())
     style = libxslt.parseStylesheetDoc(styledoc)
     doc = libxml2.parseDoc(source_xml)
     result = style.applyStylesheet(doc, None)
-    outfile = StringIO.StringIO()
-    fn = tempfile.mktemp()
-    style.saveResultToFilename(fn, result, 0)
-    style.freeStylesheet()
-    doc.freeDoc()
-    result.freeDoc()
+    
+    result_string = ''
+    try:
+        if not RENDERVIAFILE:
+            result_string = style.saveResultToString(result)
+        else:
+            # this is how we have to do things prior to the
+            # 1.0.21 release of libxslt, a pretty horrible hack
+            fn = tempfile.mktemp()
+            style.saveResultToFilename(fn, result, 0)
 
-    f = file(fn)
-    result = f.read()
-    f.close()
-    os.remove(fn)
+            f = file(fn)
+            result_string = f.read()
+            f.close()
+            os.remove(fn)
+    finally:
+        style.freeStylesheet()
+        doc.freeDoc()
+        result.freeDoc()
 
-    return result
+    return result_string
 
 def register_plugin(plugin_registry):
     plugin_registry['http://iopen.co.nz/plugins/xslt/libxslt'] = ('libxslt',
